Fix action unit testing
Using the rewire module, we now don't have to export our hidden non-async actions to test them, we can get them using rewire. Makes for a much less error-prone application writing, it's a bit more overhead for the unit tests writing, thoughI think that's worth it.

Also see https://stackoverflow.com/questions/14874208/how-to-access-and-test-an-internal-non-exports-function-in-a-node-js-module

Closes #20